requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
anthropic>=0.18.0
python-dateutil>=2.8.2
pyyaml>=6.0
plyer>=2.1.0
schedule>=1.2.0
praw>=7.7.0
requests>=2.31.0
feedparser>=6.0.10
python-dotenv>=1.0.0